Move view models to shared library
