+++ app/src/main/java/io/reark/rxgithubapp/advanced/fragments/RepositoriesFragment.java
@@ -38,7 +38,7 @@ import io.reark.rxgithubapp.advanced.RxGitHubApp;
 import io.reark.rxgithubapp.advanced.activities.ChooseRepositoryActivity;
 import io.reark.rxgithubapp.shared.utils.ApplicationInstrumentation;
 import io.reark.rxgithubapp.advanced.view.RepositoriesView;
-import io.reark.rxgithubapp.advanced.viewmodels.RepositoriesViewModel;
+import io.reark.rxgithubapp.shared.viewmodels.RepositoriesViewModel;
 
 public class RepositoriesFragment extends Fragment {
     private RepositoriesView.ViewBinder repositoriesViewBinder;
