+++ app/src/main/java/io/reark/rxgithubapp/advanced/view/RepositoriesView.java
@@ -44,8 +44,8 @@ import io.reark.reark.utils.Preconditions;
 import io.reark.reark.utils.RxViewBinder;
 import io.reark.rxgithubapp.R;
 import io.reark.rxgithubapp.shared.pojo.GitHubRepository;
-import io.reark.rxgithubapp.advanced.viewmodels.RepositoriesViewModel;
-import io.reark.rxgithubapp.advanced.viewmodels.RepositoriesViewModel.ProgressStatus;
+import io.reark.rxgithubapp.shared.viewmodels.RepositoriesViewModel;
+import io.reark.rxgithubapp.shared.viewmodels.RepositoriesViewModel.ProgressStatus;
 import rx.Observable;
 import rx.android.schedulers.AndroidSchedulers;
 import rx.subscriptions.CompositeSubscription;
