+++ appshared/src/test/java/io/reark/rxgithubapp/shared/viewmodels/RepositoryViewModelTest.java
@@ -23,13 +23,14 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package io.reark.rxgithubapp.advanced.viewmodels;
+package io.reark.rxgithubapp.shared.viewmodels;
 
 import org.junit.Test;
 
 import io.reark.rxgithubapp.shared.pojo.GitHubOwner;
 import io.reark.rxgithubapp.shared.pojo.GitHubRepository;
 import io.reark.rxgithubapp.shared.pojo.UserSettings;
+import io.reark.rxgithubapp.shared.viewmodels.RepositoryViewModel;
 import rx.Observable;
 import rx.observers.TestSubscriber;
